moved exceptions from product name too
